Move FilterPopover to it's own subfolder
