+++ ionic2/app/pages/map/filterpopover/filterpopover.scss
